Re-shuffle some code, remove unused code, and add more relations/polish things up
