+++ server/testdata/task.json
@@ -1,15 +1,13 @@
 {
     "seekerID": 1,
+    "favorTypeID": "FixComputer",
     "helperID": 1,
-    "spec": {
-        "favorType": "GoShopping",
-        "instructions": "Hello guy!",
+    "request": {
+        "instructions": "Fix my computer please!",
         "seekerLocation": {
             "lat": 1,
             "lng": 2
         }
     },
-    "status": {
-        
-    }
+    "response": {}
 }
